+++ examples/expression/src/expression-lexer.ts
