+++ examples/expression/test/expression-lexer.test.ts
@@ -4,7 +4,7 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 import { describe, expect, test } from 'vitest';
-import { tokenize, TokenType } from '../src/lexer.js';
+import { tokenize, TokenType } from '../src/expression-lexer.js';
 
 function expectTokenTypes(text: string, ...expecteds: TokenType[]) {
     const actuals = [...tokenize(text)].map(t => t.type);
